requirements.txt
nats-py==2.8.0
pyaes==1.6.1
pyasn1==0.6.0
python-decouple==3.8
rsa==4.9
Telethon==1.36.0
